more refactoring; noting that test_vis passes
